requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
xgboost
kafka-python
redis
orjson
loguru
prometheus-client
pydantic-settings
httpx
pyarrow
rich
pyyaml
ujson
grpcio
grpcio-tools
gunicorn
pytest
